[sharded_tensor] refactor validation utils for shardedmetadata
Differential Revision: [D29872128](https://our.internmc.facebook.com/intern/diff/D29872128/)

[ghstack-poisoned]
